Moved Connection struct to NEAT class. Removed WeightHandler and LayerManager class because it is not needed anymore.
