+++ EasyNNFramework/NEAT/NEAT.cs
@@ -369,4 +369,26 @@ namespace EasyNNFramework.NEAT {
             return highestLayer + 1;
         }
     }
+
+    [Serializable]
+    public struct Connection : IEquatable<Connection> {
+        public float weight;
+        public int targetID, sourceID;
+
+        public Connection(int sourceID, int targetID, float weight) {
+            this.sourceID = sourceID;
+            this.targetID = targetID;
+            this.weight = weight;
+        }
+
+        public override bool Equals(object obj) => obj is Connection n && Equals(n);
+
+        public static bool operator ==(Connection lf, Connection ri) => lf.Equals(ri);
+
+        public static bool operator !=(Connection lf, Connection ri) => !(lf == ri);
+
+        public override int GetHashCode() => (sourceID.GetHashCode() + targetID.GetHashCode() + weight.GetHashCode()).GetHashCode();
+
+        public bool Equals(Connection obj) => obj.sourceID == sourceID && obj.targetID == targetID;
+    }
 }
